Refactor ACM package with the deployment
